common/types/form.tsのstack関連の型を移動
