+++ features/stacks/types/api.ts
@@ -1,10 +1,10 @@
 import { ApiOptions } from "@/common/types/api";
 import { SessionData } from "@/features/sessions/types/session";
 import { SetErrorMessages } from "@/common/types/validator";
-import { StackFormDataParams } from "@/common/types/form";
+import { StackFormDataParams } from "./context";
 
 export type CreateStackProps = SetErrorMessages & {
   options: ApiOptions;
   sessionData: SessionData;
   stackFormData: StackFormDataParams;
-}
+};
