+++ features/stacks/types/context.ts
@@ -1,5 +1,12 @@
-import { StackFormDataParams } from "@/common/types/form";
 import { Dispatch, SetStateAction } from "react";
+
+export type StackFormDataParams = {
+  skill: string;
+  stacked_at: Date;
+  minutes: number;
+  title: string;
+  description: string;
+};
 
 export type StackFormContextProps = {
   stackFormData: StackFormDataParams;
